pill_recognition_ai/requirements.txt
# 알약 인식 AI 모델 개발을 위한 의존성 패키지 (RTX 3060Ti 최적화)

# 핵심 ML/AI 라이브러리
ultralytics>=8.0.0  # YOLOv8
torch>=2.0.0+cu118  # CUDA 11.8 지원
torchvision>=0.15.0+cu118
opencv-python>=4.8.0
Pillow>=9.5.0

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 이미지 처리 및 시각화
matplotlib>=3.7.0
seaborn>=0.12.0
albumentations>=1.3.0

# 모델 최적화 (RTX 3060Ti 최적화)
onnx>=1.14.0
onnxruntime-gpu>=1.15.0  # GPU 가속 ONNX Runtime
# tensorrt>=8.6.0  # 필요시 별도 설치

# GPU 모니터링
GPUtil>=1.4.0
psutil>=5.9.0

# 유틸리티
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.15.0  # 실험 추적
tensorboard>=2.13.0

# 개발 도구
jupyter>=1.0.0
ipykernel>=6.23.0

# 메모리 최적화
memory-profiler>=0.60.0
